requirements.txt
aiohttp
python-dotenv
tqdm